requirements.txt
Django==5.0.4
django-environ==0.11.2
django-session-security==2.6.7
django-import-export==4.1.1
Pillow>=10.4.0
python-dotenv==1.0.1